+++ deepmd/dpmodel/model/linear_atomic_model.py
@@ -7,13 +7,14 @@ from typing import (
     Union,
 )
 
-import torch
+import numpy as np
+from abc import abstractmethod
 
 from deepmd.dpmodel import (
     FittingOutputDef,
     OutputVariableDef,
 )
-from deepmd.pt.utils.nlist import (
+from deepmd.dpmodel.utils.nlist import (
     build_multiple_neighbor_list,
     get_multiple_nlist_key,
     nlist_distinguish_types,
@@ -25,37 +26,26 @@ from .base_atomic_model import (
 from .dp_atomic_model import (
     DPAtomicModel,
 )
-from .model import (
-    BaseModel,
-)
-from .pair_tab_model import (
+from .pairtab_atomic_model import (
     PairTabModel,
 )
 
-
-class LinearModel(BaseModel, BaseAtomicModel):
+class LinearAtomicModel(BaseAtomicModel):
     """Linear model make linear combinations of several existing models.
 
     Parameters
     ----------
     models : list[DPAtomicModel or PairTabModel]
         A list of models to be combined. PairTabModel must be used together with a DPAtomicModel.
-    weights : list[float] or str
-        If the type is list[float], a list of weights for each model.
-        If "mean", the weights are set to be 1 / len(models).
-        If "sum", the weights are set to be 1.
-        If "switch_by_softmin_pair_distance", the weights are calculated internally. This only allows the combination of a PairTabModel and a DPAtomicModel.
     """
 
     def __init__(
         self,
         models: List[BaseAtomicModel],
-        weights: Union[str, List[float]],
         **kwargs,
     ):
         super().__init__()
         self.models = models
-        self.weights = weights
         self.distinguish_type_list = [
             model.distinguish_types() for model in self.models
         ]
@@ -73,7 +63,6 @@ class LinearModel(BaseModel, BaseAtomicModel):
         return [model.get_rcut() for model in self.models]
 
     def get_sel(self) -> List[int]:
-        """Get the processed sels for each individual models. Not distinguishing types."""
         return [model.get_nsel() for model in self.models]
 
     def get_original_sels(self) -> List[Union[int, List[int]]]:
@@ -92,10 +81,10 @@ class LinearModel(BaseModel, BaseAtomicModel):
         extended_coord,
         extended_atype,
         nlist,
-        mapping: Optional[torch.Tensor] = None,
-        fparam: Optional[torch.Tensor] = None,
-        aparam: Optional[torch.Tensor] = None,
-    ) -> Dict[str, torch.Tensor]:
+        mapping: Optional[np.ndarray] = None,
+        fparam: Optional[np.ndarray] = None,
+        aparam: Optional[np.ndarray] = None,
+    ) -> Dict[str, np.ndarray]:
         """Return atomic prediction.
 
         Parameters
@@ -119,12 +108,10 @@ class LinearModel(BaseModel, BaseAtomicModel):
             the result dict, defined by the fitting net output def.
         """
         nframes, nloc, nnei = nlist.shape
-        if self.do_grad():
-            extended_coord.requires_grad_(True)
-        self.extended_coord = extended_coord.view(nframes, -1, 3)
+        self.extended_coord = extended_coord.reshape(nframes, -1, 3)
         sorted_rcuts, sorted_sels = self._sort_rcuts_sels()
         nlists = build_multiple_neighbor_list(
-            extended_coord,
+            self.extended_coord,
             nlist,
             sorted_rcuts,
             sorted_sels,
@@ -139,7 +126,6 @@ class LinearModel(BaseModel, BaseAtomicModel):
                 self.distinguish_type_list, raw_nlists, self.get_original_sels()
             )
         ]
-
         ener_list = [
             model.forward_atomic(
                 self.extended_coord,
@@ -149,14 +135,13 @@ class LinearModel(BaseModel, BaseAtomicModel):
             )["energy"]
             for model, nl in zip(self.models, self.nlists_)
         ]
-
-        weights = self._compute_weight()
+        self.weights = self._compute_weight()
         self.atomic_bias = None
         if self.atomic_bias is not None:
             raise NotImplementedError("Need to add bias in a future PR.")
         else:
             fit_ret = {
-                "energy": sum(torch.stack(ener_list) * torch.stack(weights)),
+                "energy": sum(np.stack(ener_list) * np.stack(self.weights)),
             }  # (nframes, nloc, 1)
         return fit_ret
 
@@ -172,38 +157,21 @@ class LinearModel(BaseModel, BaseAtomicModel):
     def serialize(self) -> dict:
         return {
             "models": [model.serialize() for model in self.models],
-            "weights": self.weights,
         }
 
     @classmethod
-    def deserialize(cls, data) -> "LinearModel":
-        weights = data["weights"]
+    def deserialize(cls, data) -> "LinearAtomicModel":
 
-        if weights == "switch_by_softmin_pair_distance":
-            raise NotImplementedError("Use ZBLAtomicModel instead of LinearModel.")
-        else:
-            models = [DPAtomicModel.deserialize(model) for model in data["models"]]
-        return cls(models, weights)
+        models = [DPAtomicModel.deserialize(model) for model in data["models"]]
+        return cls(models)
 
-    def _compute_weight(self) -> torch.Tensor:
-        if isinstance(self.weights, list):
-            if len(self.weights) != len(self.models):
-                raise ValueError(
-                    "The length of weights is not equal to the number of models"
-                )
-            return self.weights
-        elif self.weights == "mean":
-            return [1 / len(self.models) for _ in range(len(self.models))]
-        elif self.weights == "sum":
-            return [1 for _ in range(len(self.models))]
-        # TODO: add more weights, for example, so-called committee models
-        elif self.weights == "switch_by_softmin_pair_distance":
-            raise NotImplementedError("Use ZBLAtomicModel instead of LinearModel.")
-        else:
-            raise ValueError(f"Invalid weights {self.weights}")
+    @abstractmethod
+    def _compute_weight(self) -> np.ndarray:
+        """This should be a list of user defined weights that matches the number of models to be combined."""
+        raise NotImplementedError
 
 
-class ZBLAtomicModel(LinearModel):
+class DPZBLLinearAtomicModel(LinearAtomicModel):
     """Model linearly combine a list of AtomicModels.
 
     Parameters
@@ -218,18 +186,13 @@ class ZBLAtomicModel(LinearModel):
         zbl_model: PairTabModel,
         sw_rmin: float,
         sw_rmax: float,
-        weights="switch_by_softmin_pair_distance",
         smin_alpha: Optional[float] = 0.1,
         **kwargs,
     ):
         models = [dp_model, zbl_model]
-        super().__init__(models, weights, **kwargs)
+        super().__init__(models, **kwargs)
         self.dp_model = dp_model
         self.zbl_model = zbl_model
-        if weights != "switch_by_softmin_pair_distance":
-            raise ValueError(
-                "ZBLAtomicModel only supports weights 'switch_by_softmin_pair_distance'."
-            )
 
         self.sw_rmin = sw_rmin
         self.sw_rmax = sw_rmax
@@ -239,41 +202,34 @@ class ZBLAtomicModel(LinearModel):
         return {
             "dp_model": self.dp_model.serialize(),
             "zbl_model": self.zbl_model.serialize(),
-            "weights": self.weights,
             "sw_rmin": self.sw_rmin,
             "sw_rmax": self.sw_rmax,
             "smin_alpha": self.smin_alpha,
         }
 
     @classmethod
-    def deserialize(cls, data) -> "ZBLAtomicModel":
-        weights = data["weights"]
+    def deserialize(cls, data) -> "DPZBLLinearAtomicModel":
         sw_rmin = data["sw_rmin"]
         sw_rmax = data["sw_rmax"]
         smin_alpha = data["smin_alpha"]
 
-        if weights == "switch_by_softmin_pair_distance":
-            dp_model = DPAtomicModel.deserialize(data["dp_model"])
-            zbl_model = PairTabModel.deserialize(data["zbl_model"])
-        else:
-            raise ValueError(
-                "ZBLAtomicModel only supports weights 'switch_by_softmin_pair_distance'."
-            )
+        dp_model = DPAtomicModel.deserialize(data["dp_model"])
+        zbl_model = PairTabModel.deserialize(data["zbl_model"])
+        
         return cls(
             dp_model=dp_model,
             zbl_model=zbl_model,
-            weights=weights,
             sw_rmin=sw_rmin,
             sw_rmax=sw_rmax,
             smin_alpha=smin_alpha,
         )
 
-    def _compute_weight(self) -> torch.Tensor:
+    def _compute_weight(self) -> List[np.ndarray]:
         """ZBL weight.
 
         Returns
         -------
-        torch.Tensor
+        List[np.ndarray]
             the atomic ZBL weight for interpolation. (nframes, nloc, 1)
         """
         assert (
@@ -289,32 +245,36 @@ class ZBLAtomicModel(LinearModel):
         # use the larger rr based on nlist
         nlist_larger = zbl_nlist if zbl_nnei >= dp_nnei else dp_nlist
         nloc = nlist_larger.shape[1]
-        masked_nlist = torch.clamp(nlist_larger, 0)
-        pairwise_rr = (
-            (self.extended_coord.unsqueeze(2) - self.extended_coord.unsqueeze(1))
-            .pow(2)
-            .sum(-1)
-            .sqrt()
+        masked_nlist = np.clip(nlist_larger, 0, None)
+        pairwise_rr = np.sqrt(
+            np.sum(
+                np.power(
+                    (
+                        np.expand_dims(self.extended_coord, 2)
+                        - np.expand_dims(self.extended_coord, 1)
+                    ),
+                    2,
+                ),
+                axis=-1,
+            )
         )
-        rr = torch.gather(
-            pairwise_rr[:, :nloc, :], 2, masked_nlist
-        )  # nframes, nloc, nnei
 
-        numerator = torch.sum(
-            rr * torch.exp(-rr / self.smin_alpha), dim=-1
+        rr = np.take_along_axis(pairwise_rr[:, :nloc, :], masked_nlist, 2)
+
+        numerator = np.sum(
+            rr * np.exp(-rr / self.smin_alpha), axis=-1
         )  # masked nnei will be zero, no need to handle
-        denominator = torch.sum(
-            torch.where(
+        denominator = np.sum(
+            np.where(
                 nlist_larger != -1,
-                torch.exp(-rr / self.smin_alpha),
-                torch.zeros_like(nlist_larger),
+                np.exp(-rr / self.smin_alpha),
+                np.zeros_like(nlist_larger),
             ),
-            dim=-1,
+            axis=-1,
         )  # handle masked nnei.
-
-        sigma = numerator / denominator  # nfrmes, nloc
+        sigma = numerator / denominator
         u = (sigma - self.sw_rmin) / (self.sw_rmax - self.sw_rmin)
-        coef = torch.zeros_like(u)
+        coef = np.zeros_like(u)
         left_mask = sigma < self.sw_rmin
         mid_mask = (self.sw_rmin <= sigma) & (sigma < self.sw_rmax)
         right_mask = sigma >= self.sw_rmax
@@ -322,5 +282,5 @@ class ZBLAtomicModel(LinearModel):
         smooth = -6 * u**5 + 15 * u**4 - 10 * u**3 + 1
         coef[mid_mask] = smooth[mid_mask]
         coef[right_mask] = 0
-        self.zbl_weight = coef  # nframes, nloc
-        return [1 - coef.unsqueeze(-1), coef.unsqueeze(-1)]  # to match the model order.
+        self.zbl_weight = coef
+        return [1 - np.expand_dims(coef, -1), np.expand_dims(coef, -1)]
